fix: Enhanced WordPress mock types and removed redundant mock file for cleaner test setup
